requirements.txt
aiofiles==24.1.0
aiohttp==3.10.5
aiohttp==3.12.13
anthropic==0.57.1
asgiref==3.8.1
azure_storage==0.37.0
backoff==2.2.1
crawl4ai==0.6.3
Django==5.2.4
django_storages==1.14.6
djangorestframework==3.16.0
drf_yasg==1.21.10
dspy==2.6.27
dspy_ai==2.4.9
duckduckgo_search==8.0.5
easyocr==1.7.2
faster_whisper==1.1.1
ffmpeg==1.4
ffmpeg_python==0.2.0
fitz==0.0.1.dev2
google_api_python_client==2.174.0
httpx==0.28.1
ImageHash==4.3.2
langchain==0.3.26
langchain_community==0.3.27
langchain_huggingface==0.3.0
langchain_milvus==0.2.1
langchain_openai==0.3.27
langchain_qdrant==0.2.0
langchain_text_splitters==0.3.8
litellm==1.74.2
markdown_pdf==1.7
marker==2.1.5
numpy==2.3.1
openai==1.95.0
pandas==2.3.1
Pillow==11.3.0
protobuf==4.25.3
pydantic==2.11.7
pymilvus==2.5.12
pymupdf==1.26.3
PyPDF2==3.0.1
python-dotenv==1.1.1
python_magic==0.4.27
pytz==2024.1
pytz==2025.2
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.2
qdrant_client==1.14.3
rank_bm25==0.2.2
redis==6.2.0
regex==2024.9.11
regex==2024.11.6
Requests==2.32.4
rq==2.4.0
scikit_learn==1.7.0
sentence_transformers==5.0.0
tavily_python==0.7.9
toml==0.10.2
torch==2.7.1
tqdm==4.66.5
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.53.0
ujson==5.10.0
yt_dlp==2025.6.30
